issue-671
`createNumberMask` integer limit doesn't work when we enter negative (minus) value.

If we have set integerLimit to 16 digit and also allowed negative value (allowNegative). If we enter value in negative it only allow to enter 14 digit instead of 16. If we enter positive value (without - sign) it allows to enter 16 digit.