Type textMaskConfig
Put types on the properties of textMaskConfig. 

This allows, for example,, the MaskedInputDirective to be extended and wrapped by a directive that will access textMaskConfig without the mediation of a template. This way errors can be avoided when assigning a boolean or a function to textMaskConfig.mask, and showMask can be assigned as well.