Add hook for parent component to be notified when input value changes
Text mask component directly modifies the input element's dom `value`.
This commit adds a hook prop that can be used to notify parent
components when the input's value changes so that they can update their
own state to reflect the new value in the dom.

This came up for me when using `MaskedInput` inside a `Formik` form.
When the `mask` prop to `MaskedInput` changes, it can change the value
of the input element and leave the value stored in `Formik` stale. With
this `onSetInputValue` hook prop, we can provide a function that syncs
the `Formik` state (via `setFieldValue`) when the `value` in the dom is
changed.

Repro here: https://codesandbox.io/embed/l7or5346wq